Question aggregate created
Modified the eval.py to include the question aggregate. This aggregate was made so that there is a single class object which can reference all required parts needed to take a question, translate it to match the dictionary, answer the question, and return it to the user in the language matching their input language.